# backend/docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: support@tesla-app.com
    name: API Support
    url: http://localhost:8080
  description: API for Tesla driving scenarios and trip management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Tesla App API
  version: "1.0"
paths:
...
  /api/trips/{trip_id}/scenarios/{scenario_id}:
    delete:
      consumes:
      - application/json
      description: Remove scenario from trip (only for draft trips by owner)
      parameters:
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      - description: Scenario ID
        in: path
        name: scenario_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Scenario removed from trip
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "403":
          description: Access denied
          schema:
            type: object
        "404":
          description: Trip not found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Remove scenario from trip
      tags:
      - TripScenarios
    put:
      consumes:
      - application/json
      description: Update scenario duration in trip (only for draft trips by owner)
      parameters:
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      - description: Scenario ID
        in: path
        name: scenario_id
        required: true
        type: integer
      - description: Duration data
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Trip scenario updated
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "403":
          description: Access denied
          schema:
            type: object
        "404":
          description: Trip not found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update trip scenario
      tags:
      - TripScenarios
...
securityDefinitions:
  BearerAuth:
    description: JWT Authorization header using the Bearer scheme
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"